requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cryptography==44.0.2
ctree==0.1.9
cycler==0.12.1
deepseek==1.0.0
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2024.12.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
gradio==5.27.0
gradio_client==1.9.0
graphviz==0.20.3
greenlet==3.2.1
groovy==0.1.2
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mem0ai==0.1.94
milvus-lite==2.4.11
monotonic==1.6
mpmath==1.3.0
narwhals==1.36.0
neo4j==5.28.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.58.1
orjson==3.10.16
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
posthog==3.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyautogen==0.9
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
PyGithub==2.6.1
PyJWT==2.10.1
pymilvus==2.5.4
Pympler==1.1
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.7
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.3.1
serpapi==0.1.5
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
termcolor==3.0.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typer==0.15.2
typing_extensions==4.13.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.2
watchdog==6.0.0
websockets==15.0.1
wrapt==1.17.2
